backend/requirements.txt
Django==4.2
gunicorn>=20.0.4

djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

psycopg2-binary==2.9.1
coverage==7.5.4

django-password-validators==1.7.3
argon2-cffi==21.1.0
bcrypt==3.2.0